STYLE: cast the official way -> hopefully no more warnings
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=4952
git-svn-id: http://svn.slicer.org/Slicer4/trunk@4952 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Modules/VolumeRendering/vtkLabelMapPiecewiseFunction.cxx
@@ -91,7 +91,7 @@ void vtkLabelMapPiecewiseFunction::Init(vtkMRMLScalarVolumeNode *node,double opa
     this->AdjustRange(lookup->GetTableRange());
     Opacities=new double[lookup->GetNumberOfTableValues()];
     this->ColorNode=node->GetVolumeDisplayNode()->GetColorNode();
-    this->Size=lookup->GetTableRange()[1]-lookup->GetTableRange()[0];
+    this->Size=(int) (lookup->GetTableRange()[1]-lookup->GetTableRange()[0]);
     int index=0;
     for (int i=(int)lookup->GetTableRange()[0];i<lookup->GetTableRange()[1];i++)
     {
